[Devtools] Card validation (#133)
* Card validation

* Fix merge conflict error

---------
